+++ auth/src/main/java/com/umc/auth/jwt/JwtFilter.java
@@ -2,6 +2,7 @@ package com.umc.auth.jwt;
 
 
 import com.umc.auth.dto.TokenDto;
+import com.umc.common.TokenValidation;
 import jakarta.servlet.FilterChain;
 import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServletRequest;
